requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.2.10           # via django
django-filter==2.3.0      # via -r requirements.in
django==3.1               # via -r requirements.in, django-filter, djangorestframework, zappa-django-utils
djangorestframework==3.11.1  # via -r requirements.in
pytz==2020.1              # via django
sqlparse==0.3.1           # via django
werkzeug==1.0.1           # via -r requirements.in
whitenoise==5.2.0         # via -r requirements.in
zappa-django-utils==0.4.1  # via -r requirements.in
